+++ Changes.md
@@ -5,6 +5,7 @@ Fixes
 -----
 
 - WidgetAlgo : Fixed issue preventing `grab()` from capturing popup menus on Windows.
+- ShowURL : Fixed opening of "file://" URLs on Windows (#5861).
 
 Documentation
 -------------